pyproject.toml
[project]
name = "notes-agent"
version = "1.0.0"
description = "MCP server for managing handwritten notes with AI-powered cleanup and search"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pathlib2>=2.3.0",
    "orjson>=3.9.0",
    "anthropic>=0.25.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0"
]
